Fix: merge 버그 수정

+++ src/main/java/com/blog/domain/user/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.blog.domain.user.repository;
 
 import com.blog.domain.user.domain.User;
+import org.apache.el.lang.ELArithmetic;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +9,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
